re-arrange libs
